dentai/requirements.txt
asgiref==3.8.1
cffi==1.17.1
colorama==0.4.6
cryptography==45.0.5
DateTime==5.5
dj-database-url==3.0.1
Django==5.2.3
django-cors-headers==4.7.0
django-extensions==4.1
django-filter==25.1
django-sslserver==0.22
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
fuzzywuzzy==0.18.0
iniconfig==2.1.0
joblib==1.5.1
Levenshtein==0.27.1
MarkupSafe==3.0.2
numpy==2.3.1
packaging==25.0
pandas==2.3.0
pluggy==1.6.0
psycopg2-binary==2.9.10
pycparser==2.22
Pygments==2.19.2
PyJWT==2.9.0
pyOpenSSL==25.1.0
pytest==8.4.1
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-Levenshtein==0.27.1
pytz==2025.2
RapidFuzz==3.13.0
rest-framework-simplejwt==0.0.2
scikit-learn==1.7.0
scipy==1.16.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
tzdata==2025.2
Werkzeug==3.1.3
zope.interface==7.2
